Home

NoiseJul.jl

A Julia package of modules for seismological processing and analysis.

Installation

From the Julia command prompt:

Press ] to enter pkg.

Type add xxx/NoiseJul.jl

Press backspace to exit pkg.

Type using NoiseJul

Note

If you have an error when adding NoiseJul, try

'using Pkg; Pkg.pkg"add https://github.com/xtyangpsp/SeisConvert.jl.git https://github.com/kura-okubo/SeisDownload.jl.git https://github.com/jaredbryan881/SeisXcorrelation.jl.git https://github.com/kura-okubo/SeisRemoveEQ.jl.git https://github.com/kura-okubo/SeisBeamforming.git";'

Note

Since it imports large number of submodules, precompile process takes a minute.

Modules

SeisIO: Julia language support for geophysical time series data

Author: jpjones76

SeisNoise: Ambient Noise Cross-Correlation in Julia

Author: Tim Clements

SeisDownload: Mass download of Seismic data with SeisIO

Author: Kurama Okubo

SeisConvert: Data convertor from SAC format

Author: Xiaotao Yang

SeisXcorrelation (under developing): High-order CC with SeisNoise

Author: Jared Bryan

SeisRemoveEQ: Removing earthquake and tremor from raw data using kurtosis & STA/LTA

Author: Kurama Okubo

SeisBeamforming: Compute Beampower for Beamforming analysis

Author: Kurama Okubo

Documents

NoiseJul.NoiseJul

NoiseJul.NoiseJul — Module.

NoiseJul

A Julia module for seismological processing and analysis. This module calls all submodules to perform seismological processing and analysis (data download, cross-correlation, noise monitoring etc.) Each document can be found in their github repository.

"IO" => "SeisIO",

"Noise" => "SeisNoise",

"Download" => "https://github.com/kura-okubo/SeisDownload",

"Convert" => "https://github.com/xtyangpsp/SeisConvert.jl.git",

#currently NOT available "Xcorrelation" => "https://github.com/jaredbryan881/SeisXcorrelation.jl",

"RemoveEQ" => "https://github.com/kura-okubo/SeisRemoveEQ.jl",

"Beamforming" => "https://github.com/kura-okubo/SeisBeamforming.jl.git")

example:

S = NoiseJul.IO.SeisData()

Copyright (c) 2019 Harvard University Earthquake Seismology Group. All Rights Reserved.

source

Copyright (c) 2019 Harvard University Earthquake Seismology Group. All Rights Reserved.